Converted the existing bot code into cogs and extensions, and moved management code into orb_control
