+++ cogs/__init__.py
@@ -1,0 +1,1 @@
+# No code, manages python imports
